Move main runner logic to library
Some of the boilerplate code that was in the wordcount example really
was going to be repeated for every single program that used the
library.  As such, it's been refactored out into the dmrgo.Main()
function which now handles all the logic for running the job.  The
main.main of the map/reduce job should now just set up the
MapReduceJob object and call the library's Main() function.