Add full map/reduce implemetation with partitioning
A full Hadoop cluster is overkill for small (<5G-10G) jobs.  These can
easily be handled by "./map |sort |./reduce", but it's nice to have all
that automated.

This commit introduces a --mapreduce parameter that will run the
mapper, paritition the output (--partitions=N), sort them, and then call
reducer.  This is still single-threaded.

This commit also fixes a small bug that was causing it to be slower than
the equivalent Python code. The original stdout writes were unbuffered,
incurring a system call overhead on each small write.  We now wrap
os.Stdout in a bufio.Writer object which gives a 2x speedup.