Initial draft of concurrent map/reduce
Using goroutines, we're now running across the input files in parallel.
The mappers partition their output into N files per input file.  After
all the mapping is done, we've generated a number of files of the form:
   tmp-map-out-<pid>-f<file_id>.<partition_id>

So, for two files and 4 partitions, we'd end up with:
   tmp-map-out-<pid>-f0.0000
   tmp-map-out-<pid>-f0.0001
   tmp-map-out-<pid>-f0.0002
   tmp-map-out-<pid>-f0.0003
   tmp-map-out-<pid>-f1.0000
   tmp-map-out-<pid>-f1.0001
   tmp-map-out-<pid>-f1.0002
   tmp-map-out-<pid>-f1.0003

The sort/reduce phase can now be done in parallel across the partition
ids.  (I.e with two mappers running, first we'd spawn a sort/reduce on
all the files with partition id 0000 and 0001, and as those phases
finished we'd proceed with partition id 0002 and then 0003).

The code is still ugly and needs to be cleaned up, but as a first draft
I'm pretty happy with how the goroutines simplified the concurrency.