First draft of tab-separated protocol.
The JSON one is too slow to be used in production.  In my tests, it's
about 15 times slower than just scalling Sscan.  This a simpler
text-based serialization format.  We write out lines of values,
separated by tabs.  For a struct or an array, we just print out the
elements/fields in order.  Otherwise, it pretty much ends up as the
single-valued string representation of the primitive type
(int/float/string).

For the unmarshal side, I split on tabs and then fill in the values in
order of the struct that was passed in.  Arrays aren't handled yet
(it's late), but should be "trivial" to add.

This almost certianly need some cleanup, but for now it appears to
work as a proof-of-concept.